Move influxdb package in main io.vertx.ext.monitoring
