+++ reporters/influxdb/log-test.properties
@@ -22,4 +22,3 @@ java.util.logging.ConsoleHandler.level=ALL
 java.util.logging.ConsoleHandler.formatter=io.vertx.core.logging.VertxLoggerFormatter
 
 #io.vertx.ext.monitoring.level=FINEST
-#io.vertx.ext.influxdb.level=FINEST
