+++ reporters/influxdb/src/main/java/io/vertx/ext/monitoring/influxdb/impl/BatchPoints.java
@@ -14,7 +14,7 @@
  * You may elect to redistribute this code under either of these licenses.
  */
 
-package io.vertx.ext.influxdb.impl;
+package io.vertx.ext.monitoring.influxdb.impl;
 
 import java.util.ArrayList;
 import java.util.Arrays;
