+++ reporters/influxdb/src/main/kotlin/io/vertx/kotlin/ext/monitoring/influxdb/AuthenticationOptions.kt
@@ -1,9 +1,9 @@
-package io.vertx.kotlin.ext.influxdb
+package io.vertx.kotlin.ext.monitoring.influxdb
 
 import io.vertx.ext.monitoring.influxdb.AuthenticationOptions
 
 /**
- * A function providing a DSL for building [io.vertx.ext.influxdb.AuthenticationOptions] objects.
+ * A function providing a DSL for building [io.vertx.ext.monitoring.influxdb.AuthenticationOptions] objects.
  *
  * Authentication options.
  *
@@ -12,12 +12,12 @@ import io.vertx.ext.monitoring.influxdb.AuthenticationOptions
  * @param username  Set the identifier used for authentication.
  *
  * <p/>
- * NOTE: This function has been automatically generated from the [AuthenticationOptions original] using Vert.x codegen.
+ * NOTE: This function has been automatically generated from the [io.vertx.ext.monitoring.influxdb.AuthenticationOptions original] using Vert.x codegen.
  */
 fun AuthenticationOptions(
   enabled: Boolean? = null,
   secret: String? = null,
-  username: String? = null): AuthenticationOptions = AuthenticationOptions().apply {
+  username: String? = null): AuthenticationOptions = io.vertx.ext.monitoring.influxdb.AuthenticationOptions().apply {
 
   if (enabled != null) {
     this.setEnabled(enabled)
