+++ reporters/influxdb/src/main/resources/META-INF/services/io.vertx.core.spi.VertxMetricsFactory
@@ -14,4 +14,4 @@
 #  You may elect to redistribute this code under either of these licenses.
 #
 
-io.vertx.ext.influxdb.impl.VertxMetricsFactoryImpl
+io.vertx.ext.monitoring.influxdb.impl.VertxMetricsFactoryImpl
